% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{estimate_lnorm}
\alias{estimate_lnorm}
\title{Estimate Log-Normal Parameters from Censored Data}
\usage{
estimate_lnorm(y, n_missing)
}
\arguments{
\item{y}{Numeric vector. The observed (untransformed) data.}

\item{n_missing}{Integer. The number of missing (censored) observations.}
}
\value{
A list containing:
\describe{
  \item{log_mean}{The estimated mean of the log-transformed data.}
  \item{log_sd}{The estimated standard deviation of the log-transformed data.}
}
}
\description{
Estimates the mean and standard deviation of the underlying log-normal distribution
from left-censored data.
}
