% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_from_sd}
\alias{impute_from_sd}
\title{Estimate Gini Coefficient from Log-Normal SD}
\usage{
impute_from_sd(y, n_missing)
}
\arguments{
\item{y}{Numeric vector. The observed (untransformed) data.}

\item{n_missing}{Integer. The expected number of missing observations.}
}
\value{
A numeric vector of length 1 containing the estimated Gini coefficient.
}
\description{
Calculates the Gini coefficient based on the estimated standard deviation
of the log-normal distribution fitted to the censored data.
}
