% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_from_subsistence}
\alias{impute_from_subsistence}
\title{Impute Missing Values with Subsistence Income}
\usage{
impute_from_subsistence(y, n_missing, y_subsistence)
}
\arguments{
\item{y}{Numeric vector. The observed (untransformed) data.}

\item{n_missing}{Integer. The expected number of missing observations.}

\item{y_subsistence}{Numeric. The subsistence income value to use for imputation.}
}
\value{
A numeric vector containing the original observed values `y` and the imputed values.
}
\description{
Fills missing values with a fixed subsistence income value.
}
